History Monitor verplaatst naar core. Mogelijk probleem met jaarwisseling gefixt. TMobile geeft nu 'gebruikersnaam'. 'Wis geschiedenis'-knop is nu naar het menu.

+++ core/src/main/java/nl/haroid/access/BalanceRepository.java
@@ -155,8 +155,10 @@ public final class BalanceRepository {
 
     private void upgrade(Storage storage, int oldVersion, int newVersion) {
         LOGGER.info("onUpgrade() "  + oldVersion + " -> " + newVersion);
-        storage.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
-        create(storage);
+        if (newVersion > oldVersion) {
+            storage.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
+            create(storage);
+        }
     }
 
     private void reset(Storage storage) {
